requirements/base.txt
Django==1.6.5
bpython==0.13
django-braces==1.4.0
django-jfu==2.0.1
django-model-utils==2.0.3
easy-thumbnails==2.0.1
logutils==0.3.3
Pillow==2.4.0
pip-tools==0.3.4
South==0.8.4
